+++ packages/atomic/cypress/integration/facets/category-facet/category-facet-assertions.ts
@@ -176,9 +176,7 @@ export function assertFirstChildContains(value: string) {
 
 export function assertNumberOfSearchResults(numberOfResults: number) {
   it(`should have ${numberOfResults} search results`, () => {
-    ResultListSelectors.shadow()
-      .find(ResultListSelectors.result)
-      .should('have.length', numberOfResults);
+    ResultListSelectors.result().should('have.length', numberOfResults);
   });
 }