+++ packages/atomic/cypress/integration/result-list/result-components/result-text.cypress.ts
@@ -1,47 +1,45 @@
 import {HighlightKeyword} from '@coveo/headless';
-import {generateComponentHTML} from '../../fixtures/test-fixture';
-import {interceptSearchResponse, setUpPage} from '../../utils/setupComponent';
-import {executeFirstSearch, setFieldCaptions} from '../search-interface-utils';
 import {
-  generateResultList,
-  generateResultTemplate,
-  getFirstResult,
-} from '../result-list/result-list-selectors';
-import {ResultTextSelectors} from './result-text-selectors';
+  generateComponentHTML,
+  TagProps,
+  TestFixture,
+} from '../../../fixtures/test-fixture';
+import {addResultList, buildTemplateWithSections} from '../result-list-actions';
+import {
+  resultTextComponent,
+  ResultTextSelectors,
+} from './result-text-selectors';
+
+interface ResultTextProps {
+  field?: string | number;
+  default?: string;
+  'should-highlight'?: string;
+}
+
+const addResultTextInResultList = (props: ResultTextProps = {}) =>
+  addResultList(
+    buildTemplateWithSections({
+      bottomMetadata: generateComponentHTML(
+        resultTextComponent,
+        props as TagProps
+      ),
+    })
+  );
 
 describe('Result Text Component', () => {
-  function setupResultTextPage(
-    props: Record<string, string | number>,
-    executeSearch = true
-  ) {
-    setUpPage(
-      generateResultList(
-        generateResultTemplate({
-          bottomMetadata: generateComponentHTML(
-            ResultTextSelectors.component,
-            props
-          ).outerHTML,
-        })
-      ),
-      executeSearch
-    );
-  }
-
-  function getFirstResultText() {
-    return getFirstResult().find(ResultTextSelectors.component);
-  }
-
   describe('when not used inside a result template', () => {
     beforeEach(() => {
-      setUpPage(generateComponentHTML(ResultTextSelectors.component).outerHTML);
+      new TestFixture()
+        .withElement(generateComponentHTML(resultTextComponent))
+        .init();
     });
 
     it.skip('should remove the component from the DOM', () => {
-      cy.get(ResultTextSelectors.component).should('not.exist');
+      cy.get(resultTextComponent).should('not.exist');
     });
 
     it('should log a console error', () => {
-      cy.get(ResultTextSelectors.component)
+      cy.get(resultTextComponent)
         .find('atomic-component-error')
         .should('exist');
     });
@@ -50,14 +48,18 @@ describe('Result Text Component', () => {
   describe('when the field does not exist for the result but the "default" prop is set', () => {
     const defaultText = 'Test default';
     beforeEach(() => {
-      setupResultTextPage({
-        default: defaultText,
-        field: 'thisfielddoesnotexist',
-      });
+      new TestFixture()
+        .with(
+          addResultTextInResultList({
+            default: defaultText,
+            field: 'thisfielddoesnotexist',
+          })
+        )
+        .init();
     });
 
     it('should render an "atomic-text" component with the default value', () => {
-      getFirstResultText()
+      ResultTextSelectors.firstInResult()
         .find('atomic-text')
         .shadow()
         .should('have.text', defaultText);
@@ -66,32 +68,28 @@ describe('Result Text Component', () => {
 
   describe('when the field does not exist for the result and the "default" prop is not set', () => {
     beforeEach(() => {
-      setupResultTextPage({
-        field: 'thisfielddoesnotexist',
-      });
+      new TestFixture()
+        .with(addResultTextInResultList({field: 'thisfielddoesnotexist'}))
+        .init();
     });
 
     it('should remove the component from the DOM', () => {
-      getFirstResultText().should('not.exist');
+      ResultTextSelectors.firstInResult().should('not.exist');
     });
   });
 
   describe('when the field value is not a string', () => {
     beforeEach(() => {
-      setupResultTextPage(
-        {
-          field: 420,
-        },
-        false
-      );
-      interceptSearchResponse((response) =>
-        response.results.forEach((result) => (result.raw['420'] = 'Abc'))
-      );
-      executeFirstSearch();
+      new TestFixture()
+        .with(addResultTextInResultList({field: 420}))
+        .withCustomResponse((response) =>
+          response.results.forEach((result) => (result.raw['420'] = 'Abc'))
+        )
+        .init();
     });
 
     it.skip('should remove the component from the DOM', () => {
-      getFirstResultText().should('not.exist');
+      ResultTextSelectors.firstInResult().should('not.exist');
     });
   });
 
@@ -108,23 +106,23 @@ describe('Result Text Component', () => {
       shouldHighlight: boolean,
       highlightsAvailable: boolean
     ) {
-      setupResultTextPage(
-        {
-          field: field,
-          'should-highlight': shouldHighlight.toString(),
-        },
-        false
-      );
-      interceptSearchResponse((response) =>
-        response.results.forEach((result) => {
-          result.raw[field] = rawValue;
-          if (highlightsAvailable) {
-            (result as any)[`${field}Highlights`] = [highlight];
-          }
-        })
-      );
-      setFieldCaptions(field, {[rawValue]: localizedValue});
-      executeFirstSearch();
+      new TestFixture()
+        .with(
+          addResultTextInResultList({
+            field: field,
+            'should-highlight': shouldHighlight.toString(),
+          })
+        )
+        .withCustomResponse((response) =>
+          response.results.forEach((result) => {
+            result.raw[field] = rawValue;
+            if (highlightsAvailable) {
+              (result as any)[`${field}Highlights`] = [highlight];
+            }
+          })
+        )
+        .withFieldCaptions(field, {[rawValue]: localizedValue})
+        .init();
     }
 
     describe('when the "shouldHighlight" prop is true and when highlights are available for the field', () => {
@@ -133,8 +131,8 @@ describe('Result Text Component', () => {
       });
 
       it('should render the highlighted text', () => {
-        getFirstResultText().should('have.text', rawValue);
-        getFirstResultText()
+        ResultTextSelectors.firstInResult().should('have.text', rawValue);
+        ResultTextSelectors.firstInResult()
           .find('[part="highlight"]')
           .should('have.text', highlightedValue);
       });
@@ -146,12 +144,14 @@ describe('Result Text Component', () => {
       });
 
       it('should render text value with localization', () => {
-        getFirstResultText().should('have.text', localizedValue);
-        getFirstResultText().find('[part="highlight"]').should('not.exist');
+        ResultTextSelectors.firstInResult().should('have.text', localizedValue);
+        ResultTextSelectors.firstInResult()
+          .find('[part="highlight"]')
+          .should('not.exist');
       });
 
       it.skip('should render an "atomic-text" component with the localized value', () => {
-        getFirstResultText()
+        ResultTextSelectors.firstInResult()
           .find('atomic-text')
           .shadow()
           .should('have.text', localizedValue);
@@ -164,14 +164,12 @@ describe('Result Text Component', () => {
       });
 
       it('should render text value with localization', () => {
-        getFirstResultText().should('have.text', localizedValue);
-        getFirstResultText()
-          .find(ResultTextSelectors.highlight)
-          .should('not.exist');
+        ResultTextSelectors.firstInResult().should('have.text', localizedValue);
+        ResultTextSelectors.highlight().should('not.exist');
       });
 
       it.skip('should render an "atomic-text" component with the localized value', () => {
-        getFirstResultText()
+        ResultTextSelectors.firstInResult()
           .find('atomic-text')
           .shadow()
           .should('have.text', localizedValue);