+++ packages/atomic/cypress/integration/result-list/result-list.cypress.ts
@@ -1,66 +1,55 @@
-import {setUpPage} from '../../utils/setupComponent';
 import {createAliasNavigation, PagerSelectors} from '../pager-selectors';
 import {withAnySectionnableResultList} from './result-list-utils';
-import {
-  ResultListSelectors,
-  generateResultList,
-  generateResultTemplate,
-} from './result-list-selectors';
+import {ResultListSelectors} from './result-list-selectors';
+import {generateComponentHTML, TestFixture} from '../../fixtures/test-fixture';
+import {addResultList, buildTemplateWithSections} from './result-list-actions';
 
 describe('Result List Component', () => {
-  function getFirstResult() {
-    return ResultListSelectors.shadow()
-      .find(ResultListSelectors.result)
-      .first()
-      .shadow();
-  }
-
   it('should load', () => {
-    setUpPage(generateResultList());
-    ResultListSelectors.shadow()
-      .find(ResultListSelectors.result)
-      .should('have.length.above', 0);
+    new TestFixture().with(addResultList()).init();
+    ResultListSelectors.result().should('have.length.above', 0);
   });
 
   describe('when no first search has yet been executed', () => {
     beforeEach(() => {
-      setUpPage(generateResultList(), false);
+      new TestFixture()
+        .withoutFirstAutomaticSearch()
+        .with(addResultList())
+        .init();
     });
 
     it('should render placeholder components', () => {
-      ResultListSelectors.shadow()
-        .find(ResultListSelectors.placeholder)
-        .should('be.visible');
+      ResultListSelectors.placeholder().should('be.visible');
     });
   });
 
   describe('when an initial search is executed', () => {
     it('should render the correct number of results', () => {
-      setUpPage(generateResultList());
-      ResultListSelectors.shadow()
-        .find(ResultListSelectors.result)
-        .should('have.length', 10);
+      new TestFixture().with(addResultList()).init();
+      ResultListSelectors.result().should('have.length', 10);
     });
   });
 
   describe('when multiple searches are executed', () => {
+    beforeEach(() => {
+      new TestFixture()
+        .with(addResultList())
+        .withElement(generateComponentHTML(PagerSelectors.pager))
+        .init();
+      createAliasNavigation();
+    });
+
     it('should update the results', () => {
       let firstResultHtml: string;
-      setUpPage(
-        `${generateResultList()}<${PagerSelectors.pager}></${
-          PagerSelectors.pager
-        }>`
-      );
-      createAliasNavigation();
 
-      getFirstResult().then((element) => {
+      ResultListSelectors.firstResult().then((element) => {
         firstResultHtml = element[0].innerHTML;
       });
 
       cy.get('@nextButton').click();
       cy.wait(500);
 
-      getFirstResult().should((element) => {
+      ResultListSelectors.firstResult().should((element) => {
         const secondResultHtml = element[0].innerHTML;
         expect(secondResultHtml).not.to.equal(firstResultHtml);
       });
@@ -71,43 +60,46 @@ describe('Result List Component', () => {
     const lineHeightSelector = '#line-height-el';
 
     function generateLineHeightElement() {
-      return `<div
-        id="${lineHeightSelector.slice(1)}"
-        style="background-color: red; width: var(--line-height, 0); height: var(--line-height, 0);"
-      ></div>`;
+      return generateComponentHTML('div', {
+        id: lineHeightSelector.slice(1),
+        style:
+          'background-color: red; width: var(--line-height, 0); height: var(--line-height, 0);',
+      });
     }
 
     before(() => {
-      setUpPage(
-        generateResultList(
-          generateResultTemplate({
-            title: generateLineHeightElement(),
-            excerpt: generateLineHeightElement(),
-            bottomMetadata: generateLineHeightElement(),
-          })
+      new TestFixture()
+        .with(
+          addResultList(
+            buildTemplateWithSections({
+              title: generateLineHeightElement(),
+              excerpt: generateLineHeightElement(),
+              bottomMetadata: generateLineHeightElement(),
+            })
+          )
         )
-      );
+        .init();
       ResultListSelectors.shadow().find('.list-wrapper:not(.placeholder)');
     });
 
     withAnySectionnableResultList(() => {
       it('should expose --line-height in the title section', () => {
-        getFirstResult()
-          .find(ResultListSelectors.sections.title)
+        ResultListSelectors.sections
+          .title()
           .find(lineHeightSelector)
           .should('be.visible');
       });
 
       it('should expose --line-height in the excerpt section', () => {
-        getFirstResult()
-          .find(ResultListSelectors.sections.excerpt)
+        ResultListSelectors.sections
+          .excerpt()
           .find(lineHeightSelector)
           .should('be.visible');
       });
 
       it('should expose --line-height in the bottom-metadata section', () => {
-        getFirstResult()
-          .find(ResultListSelectors.sections.bottomMetadata)
+        ResultListSelectors.sections
+          .bottomMetadata()
           .find(lineHeightSelector)
           .should('be.visible');
       });
@@ -116,24 +108,29 @@ describe('Result List Component', () => {
 
   describe('with a full result template', () => {
     function generateSimpleTextElement() {
-      return '<span>I will not use meaningless placeholder text for testing</span>';
+      const element = generateComponentHTML('span');
+      element.innerText =
+        'I will not use meaningless placeholder text for testing';
+      return element;
     }
 
     before(() => {
-      setUpPage(
-        generateResultList(
-          generateResultTemplate({
-            visual: generateSimpleTextElement(),
-            badges: generateSimpleTextElement(),
-            actions: generateSimpleTextElement(),
-            title: generateSimpleTextElement(),
-            titleMetadata: generateSimpleTextElement(),
-            emphasized: generateSimpleTextElement(),
-            excerpt: generateSimpleTextElement(),
-            bottomMetadata: generateSimpleTextElement(),
-          })
+      new TestFixture()
+        .with(
+          addResultList(
+            buildTemplateWithSections({
+              visual: generateSimpleTextElement(),
+              badges: generateSimpleTextElement(),
+              actions: generateSimpleTextElement(),
+              title: generateSimpleTextElement(),
+              titleMetadata: generateSimpleTextElement(),
+              emphasized: generateSimpleTextElement(),
+              excerpt: generateSimpleTextElement(),
+              bottomMetadata: generateSimpleTextElement(),
+            })
+          )
         )
-      );
+        .init();
     });
 
     withAnySectionnableResultList(() => {