+++ packages/quantic/cypress/integration/facets/facet/facet-actions.ts
@@ -1,6 +1,7 @@
 import {BaseFacetSelector} from '../facet-common-selectors';
 
 const valueCheckbox = 'input[type="checkbox"]';
+const valueLink = '.facet__value-text';
 
 export function checkFirstValue(selector: BaseFacetSelector) {
   selector.values().first().find(valueCheckbox).check({force: true});
@@ -8,4 +9,12 @@ export function checkFirstValue(selector: BaseFacetSelector) {
 
 export function checkLastValue(selector: BaseFacetSelector) {
   selector.values().last().find(valueCheckbox).check({force: true});
+}
+
+export function selectFirstLinkValue(selector: BaseFacetSelector) {
+  selector.values().first().find(valueLink).click({force: true});
+}
+
+export function selectLastLinkValue(selector: BaseFacetSelector) {
+  selector.values().last().find(valueLink).click({force: true});
 }