+++ packages/quantic/force-app/main/default/lwc/quanticRecentQueriesList/quanticRecentQueriesList.js
@@ -24,13 +24,21 @@ export default class QuanticRecentQueriesList extends LightningElement {
   @api maxLength = 10;
   /** @type {string} */
   @api engineId;
+  /** @type {boolean} */
+  @api get isCollapsed() {
+    return this._isCollapsed;
+  }
+  set isCollapsed(collapsed) {
+    this._isCollapsed = collapsed;
+  }
+    
+  /** @type {boolean} */
+  _isCollapsed = false;
   
   /** @type {import("coveo").RecentQueriesList} */
   recentQueriesList;
   /** @type {()=> void} */
   unsubscribe;
-  /** @type {boolean} */
-  isExpanded = true;
 
   connectedCallback() {
     registerComponentForInit(this, this.engineId);
@@ -69,7 +77,7 @@ export default class QuanticRecentQueriesList extends LightningElement {
   }
 
   toggleVisibility() {
-    this.isExpanded = !this.isExpanded;
+    this._isCollapsed = !this.isCollapsed;
   }
 
   get queries() {
@@ -85,11 +93,11 @@ export default class QuanticRecentQueriesList extends LightningElement {
   }
 
   get actionButtonIcon() {
-    return this.isExpanded ? 'utility:dash' : 'utility:add';
+    return this.isCollapsed ? 'utility:add' : 'utility:dash';
   }
 
   get actionButtonLabel() {
-    const label = this.isExpanded ? this.labels.collapse : this.labels.expand;
+    const label = this.isCollapsed ? this.labels.expand : this.labels.collapse;
     return I18nUtils.format(label, this.label);
   }
 }