+++ packages/quantic/force-app/main/default/lwc/quanticRecentResultsList/quanticRecentResultsList.js
@@ -27,9 +27,17 @@ export default class QuanticRecentResultsList extends LightningElement {
   @api maxLength = 10;
   /** @type {string} */
   @api label = this.labels.recentResultsLabel;
-
   /** @type {boolean} */
-  isExpanded = true;
+  @api get isCollapsed() {
+    return this._isCollapsed;
+  }
+  set isCollapsed(collapsed) {
+    this._isCollapsed = collapsed;
+  }
+    
+  /** @type {boolean} */
+  _isCollapsed = false;
+
   /** @type {import("coveo").RecentResultsList} */
   recentResultsList;
 
@@ -69,7 +77,7 @@ export default class QuanticRecentResultsList extends LightningElement {
   }
 
   toggleVisibility() {
-    this.isExpanded = !this.isExpanded;
+    this._isCollapsed = !this.isCollapsed;
   }
 
   get results() {
@@ -85,11 +93,11 @@ export default class QuanticRecentResultsList extends LightningElement {
   }
 
   get actionButtonIcon() {
-    return this.isExpanded ? 'utility:dash' : 'utility:add';
+    return this.isCollapsed ? 'utility:add' : 'utility:dash';
   }
 
   get actionButtonLabel() {
-    const label = this.isExpanded ? this.labels.collapse : this.labels.expand;
+    const label = this.isCollapsed ? this.labels.expand : this.labels.collapse;
     return I18nUtils.format(label, this.label);
   }
 }