+++ packages/quantic/force-app/main/default/lwc/quanticResultsPerPage/quanticResultsPerPage.js
@@ -17,8 +17,15 @@ export default class QuanticResultsPerPage extends LightningElement {
 
   /** @type {string} */
   @api engineId;
+  /** @type {string} */
+  @api choicesDisplayed = '10,25,50,100';
+  /** @type {number} */
+  @api initialChoice = 10;
+
   /** @type {boolean}*/
   @track hasResults
+  /** @type {number[]} */
+  @track choices = [];
   
   connectedCallback() {
     registerComponentForInit(this, this.engineId);
@@ -32,7 +39,11 @@ export default class QuanticResultsPerPage extends LightningElement {
    * @param {import("coveo").SearchEngine} engine
    */
   initialize = (engine) => {
-    this.resultsPerPage = CoveoHeadless.buildResultsPerPage(engine);
+    this.choices = this.parseChoicesDisplayed();
+    this.validateInitialChoice();
+    this.resultsPerPage = CoveoHeadless.buildResultsPerPage(engine, {
+      initialState: {numberOfResults: Number(this.initialChoice) ?? this.choices[0]},
+    });
     this.searchStatus = CoveoHeadless.buildSearchStatus(engine);
     this.unsubscribe = this.resultsPerPage.subscribe(() => this.updateState());
     this.unsubscribeSearchStatus = this.searchStatus.subscribe(() => this.updateState());
@@ -48,12 +59,26 @@ export default class QuanticResultsPerPage extends LightningElement {
     this.hasResults = this.searchStatus.state.hasResults;
   }
 
+  parseChoicesDisplayed() {
+    return this.choicesDisplayed.split(',').map((choice) => {
+      const parsedChoice = parseInt(choice, 10);
+      if (isNaN(parsedChoice)) {
+        throw new Error(`The choice value "${choice}" from the "choicesDisplayed" option is not a number.`);
+      }
+      return parsedChoice;
+    });
+  }
+
+  validateInitialChoice() {
+    if (!this.choices.includes(Number(this.initialChoice))) {
+      throw new Error(`The "initialChoice" option value "${this.initialChoice}" is not included in the "choicesDisplayed" option "${this.choicesDisplayed}".`);
+    }
+  }
+
   /**
    * @param {CustomEvent<number>} event
    */
   goto(event) {
     this.resultsPerPage.set(event.detail);
   }
-
-  resultsPerPageOptions = [10, 25, 50, 100];
 }