+++ packages/quantic/CHANGELOG.md
@@ -3,6 +3,45 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.66.0](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.65.0...@coveo/quantic@0.66.0) (2021-09-29)
+
+
+### Features
+
+* **quantic:** add apache 2.0 license ([#1270](https://github.com/coveo/ui-kit/issues/1270)) ([317ce6a](https://github.com/coveo/ui-kit/commit/317ce6a8523c7acd4eed2b52cff85c85f115ff62))
+
+
+
+
+
+# [0.65.0](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.64.8...@coveo/quantic@0.65.0) (2021-09-29)
+
+
+### Features
+
+* **quantic:** added some missing component options ([#1256](https://github.com/coveo/ui-kit/issues/1256)) ([207e3e4](https://github.com/coveo/ui-kit/commit/207e3e4d0d499386b63cb2e6c0ff24c841d353f2))
+* **quantic:** facet onhover style and display as link ([#1253](https://github.com/coveo/ui-kit/issues/1253)) ([2cb9e99](https://github.com/coveo/ui-kit/commit/2cb9e993c417a5475715bc948cf63812d46ba901))
+
+
+
+
+
+## [0.64.8](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.64.7...@coveo/quantic@0.64.8) (2021-09-28)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [0.64.7](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.64.6...@coveo/quantic@0.64.7) (2021-09-28)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
 ## [0.64.6](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.64.5...@coveo/quantic@0.64.6) (2021-09-28)